Extract open-generic-derived unification core to RoslynExtensions
The per-ancestor unification, canonical-substitution consistency check,
constraint-equivalence loop, and closure construction are pure type-theory
logic; they don't depend on parser state, [JsonDerivedType] attribute
parsing, or diagnostic emission. Move them into
RoslynExtensions.TryResolveOpenGenericDerivedType (an extension on
Compilation).

JsonSourceGenerator.Parser.TryResolveDerivedType retains the four-case
open/closed x generic/non-generic dispatch (because the closed-derived /
non-generic-base early-outs are parser-shaped) and delegates the open +
generic case to the new helper. The reflection-side resolver's MIRRORS
cross-reference is updated to point at the new location (and was already
stale -- it still named the old TryResolveOpenGenericDerivedType in
Parser, which had since been renamed to TryResolveDerivedType).

No behavior change. All 27 polymorphism / open-generic / derived-type
source-gen unit tests still pass.
