+++ src/libraries/System.Text.Json/gen/JsonSourceGenerator.Parser.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
 using System.Linq;
 using System.Text.Json.Serialization;
 using System.Threading;
@@ -1049,44 +1048,25 @@ namespace System.Text.Json.SourceGeneration
 
             /// <summary>
             /// Source-gen-side resolver for a <c>[JsonDerivedType]</c> registration against
-            /// <paramref name="baseType"/>. Handles four cases:
+            /// <paramref name="baseType"/>. Dispatches on the four open/closed × generic/non-generic
+            /// shapes:
             /// <list type="bullet">
-            ///   <item>Open derived + generic base: validates universal applicability and
-            ///   produces the closed derived type for the current closure.</item>
+            ///   <item>Open derived + generic base: delegates to
+            ///   <see cref="RoslynExtensions.TryResolveOpenGenericDerivedType"/> for
+            ///   universal-applicability validation and closure construction.</item>
             ///   <item>Open derived + non-generic base: rejected as unassignable.</item>
-            ///   <item>Closed derived + generic base: rejected. The attribute is shared across
-            ///   every closure of the base definition; a closed derived necessarily pins one
-            ///   specialization and would silently work for that closure and break for the
-            ///   others.</item>
-            ///   <item>Closed derived + non-generic base: passes through unchanged; the normal
-            ///   PolymorphicTypeResolver assignability check applies downstream.</item>
+            ///   <item>Closed derived + generic base: rejected. The attribute lives on the open
+            ///   base definition and is shared across every closure; a closed derived necessarily
+            ///   pins one specialization and would silently work for that closure and break for
+            ///   the others.</item>
+            ///   <item>Closed derived + non-generic base: passes through unchanged; the runtime
+            ///   <c>PolymorphicTypeResolver</c> assignability check applies downstream.</item>
             /// </list>
             ///
-            /// "Universal" (open + generic case) means: there is a single canonical substitution
-            /// mapping each derived type parameter to a base type parameter that simultaneously
-            /// satisfies every matching ancestor of the derived type, with every derived
-            /// constraint implied by (i.e. weaker-than-or-equal-to) the constraints on the
-            /// corresponding base parameter. Registrations that pin a particular specialization
-            /// (e.g. <c>Derived&lt;T&gt; : Base&lt;T, int&gt;</c>) are rejected: such registrations
-            /// would silently work for one base specialization and break for another, which we
-            /// treat as a misregistration regardless of which specialization is currently being
-            /// generated.
-            ///
-            /// Returns <c>true</c> with <paramref name="resolvedDerivedType"/> set (either the
-            /// closed derived type from unification, or <paramref name="declaredDerived"/> as-is
-            /// for the pass-through case). Returns <c>false</c> with a localized
+            /// Returns <see langword="true"/> with <paramref name="resolvedDerivedType"/> set
+            /// (either the closed derived type from unification, or <paramref name="declaredDerived"/>
+            /// as-is for the pass-through case). Returns <see langword="false"/> with a localized
             /// <paramref name="failureReason"/> suitable for inclusion in <c>SYSLIB1229</c>.
-            ///
-            /// IMPORTANT: The open-derived unification path (per-ancestor unification, canonical-
-            /// substitution consistency check, constraint-equivalence rules) MIRRORS the
-            /// reflection resolver
-            /// <c>DefaultJsonTypeInfoResolver.Helpers.TryResolveOpenGenericDerivedType</c> in
-            /// src/System/Text/Json/Serialization/Metadata/DefaultJsonTypeInfoResolver.Helpers.cs
-            /// and must be kept in lockstep so that source-gen and reflection produce the same
-            /// closed type for the same registration. The reflection-side wrapper splits the
-            /// four-case dispatch differently (it pre-splits the base into definition + args and
-            /// inlines the closed-derived / non-generic-base checks at the caller for caching);
-            /// the unification core is what's mirrored, not the dispatch shape.
             /// </summary>
             private bool TryResolveDerivedType(
                 ITypeSymbol declaredDerived,
@@ -1135,162 +1115,15 @@ namespace System.Text.Json.SourceGeneration
                     return false;
                 }
 
-                INamedTypeSymbol derivedDefinition = unboundDerived.OriginalDefinition;
-                INamedTypeSymbol baseDefinition = constructedBase.OriginalDefinition;
-
-                // Every ancestor of the derived type definition whose original definition matches
-                // the base type definition. For classes there is at most one such ancestor; for
-                // interfaces a derived type can implement the same interface definition multiple
-                // times with different type arguments.
-                List<INamedTypeSymbol> matchingBases = derivedDefinition
-                    .GetCompatibleGenericBaseTypes(baseDefinition)
-                    .ToList();
-
-                if (matchingBases.Count == 0)
+                if (_knownSymbols.Compilation.TryResolveOpenGenericDerivedType(
+                        unboundDerived, constructedBase,
+                        out INamedTypeSymbol? closedDerivedType, out failureReason))
                 {
-                    failureReason = SR.Polymorphism_OpenGeneric_Reason_NotAssignable;
-                    return false;
-                }
-
-                // The complete set of derived parameters that must be bound (enclosing + leaf).
-                List<ITypeParameterSymbol> requiredDerivedParams = derivedDefinition.GetAllTypeParameters();
-                List<ITypeParameterSymbol> baseParams = baseDefinition.GetAllTypeParameters();
-                var baseParamSet = new HashSet<ITypeParameterSymbol>(baseParams, SymbolEqualityComparer.Default);
-
-                // Per-ancestor independent substitutions; the universal answer must be a single
-                // canonical substitution agreed upon by every ancestor.
-                Dictionary<ITypeParameterSymbol, ITypeSymbol>? canonical = null;
-
-                foreach (INamedTypeSymbol ancestor in matchingBases)
-                {
-                    var substitution = new Dictionary<ITypeParameterSymbol, ITypeSymbol>(
-                        requiredDerivedParams.Count, SymbolEqualityComparer.Default);
-
-                    if (!ancestor.TryUnifyWith(baseDefinition, substitution))
-                    {
-                        // Some position pins a concrete type (e.g. Base<T, int>) or a constructed
-                        // pattern (e.g. Base<List<T>>) that cannot match a free base parameter.
-                        failureReason = SR.Polymorphism_OpenGeneric_Reason_NonUniversalPinning;
-                        return false;
-                    }
-
-                    foreach (ITypeParameterSymbol p in requiredDerivedParams)
-                    {
-                        if (!substitution.TryGetValue(p, out ITypeSymbol? mapped))
-                        {
-                            // E.g. D<U1, U2> : IBase<U1> -- U2 is not bound by this ancestor.
-                            failureReason = string.Format(
-                                CultureInfo.InvariantCulture,
-                                SR.Polymorphism_OpenGeneric_Reason_UnboundParameter,
-                                p.Name);
-                            return false;
-                        }
-
-                        if (mapped is not ITypeParameterSymbol mappedBaseParam || !baseParamSet.Contains(mappedBaseParam))
-                        {
-                            // Substitution value isn't one of the base's own type parameters --
-                            // happens when a derived ancestor binds a parameter to something
-                            // structural (which TryUnifyWith would have rejected) or pinned.
-                            // Treated as non-universal.
-                            failureReason = SR.Polymorphism_OpenGeneric_Reason_NonUniversalPinning;
-                            return false;
-                        }
-                    }
-
-                    if (canonical is null)
-                    {
-                        canonical = substitution;
-                    }
-                    else if (!SubstitutionsEqual(canonical, substitution))
-                    {
-                        // Two ancestors agree on independent bindings but produce different
-                        // (derived -> base) mappings, e.g. D<U1, U2> : IBase<U1, U2>, IBase<U2, U1>.
-                        // There is no single canonical answer for an arbitrary base closure.
-                        failureReason = SR.Polymorphism_OpenGeneric_Reason_AmbiguousMatch;
-                        return false;
-                    }
-                }
-
-                Debug.Assert(canonical is not null);
-
-                // Constraint equivalence: every derived parameter's constraints must exactly
-                // match the constraints on the mapped base parameter (after substitution) so
-                // that any valid closure of the base also yields a valid closure of the
-                // derived. See ReflectionExtensions.AreConstraintsEquivalent for the rationale
-                // behind exact match (vs one-sided subsumption).
-                foreach (ITypeParameterSymbol derivedParam in requiredDerivedParams)
-                {
-                    var mappedBaseParam = (ITypeParameterSymbol)canonical[derivedParam];
-                    if (!_knownSymbols.Compilation.AreConstraintsEquivalent(derivedParam, mappedBaseParam, canonical))
-                    {
-                        failureReason = string.Format(
-                            CultureInfo.InvariantCulture,
-                            SR.Polymorphism_OpenGeneric_Reason_ConstraintMismatch,
-                            derivedParam.Name,
-                            mappedBaseParam.Name);
-                        return false;
-                    }
-                }
-
-                // Closure construction: substitute the canonical mapping then specialize each
-                // base parameter to the actual closed-base type argument.
-                var baseParamPosition = new Dictionary<ITypeParameterSymbol, int>(SymbolEqualityComparer.Default);
-                for (int i = 0; i < baseParams.Count; i++)
-                {
-                    baseParamPosition[baseParams[i]] = i;
-                }
-
-                List<ITypeSymbol> constructedBaseAllArgs = GetAllTypeArguments(constructedBase);
-
-                var closedArgs = new ITypeSymbol[requiredDerivedParams.Count];
-                for (int i = 0; i < requiredDerivedParams.Count; i++)
-                {
-                    var mappedBaseParam = (ITypeParameterSymbol)canonical[requiredDerivedParams[i]];
-                    closedArgs[i] = constructedBaseAllArgs[baseParamPosition[mappedBaseParam]];
-                }
-
-                resolvedDerivedType = derivedDefinition.ConstructWithEnclosingTypeArguments(closedArgs);
-                return true;
-
-                static bool SubstitutionsEqual(
-                    Dictionary<ITypeParameterSymbol, ITypeSymbol> a,
-                    Dictionary<ITypeParameterSymbol, ITypeSymbol> b)
-                {
-                    if (a.Count != b.Count)
-                    {
-                        return false;
-                    }
-
-                    foreach (KeyValuePair<ITypeParameterSymbol, ITypeSymbol> kvp in a)
-                    {
-                        if (!b.TryGetValue(kvp.Key, out ITypeSymbol? otherValue) ||
-                            !SymbolEqualityComparer.Default.Equals(kvp.Value, otherValue))
-                        {
-                            return false;
-                        }
-                    }
-
+                    resolvedDerivedType = closedDerivedType;
                     return true;
                 }
 
-                static List<ITypeSymbol> GetAllTypeArguments(INamedTypeSymbol type)
-                {
-                    var result = new List<ITypeSymbol>();
-                    Append(type.ContainingType, result);
-                    result.AddRange(type.TypeArguments);
-                    return result;
-
-                    static void Append(INamedTypeSymbol? enclosing, List<ITypeSymbol> list)
-                    {
-                        if (enclosing is null)
-                        {
-                            return;
-                        }
-
-                        Append(enclosing.ContainingType, list);
-                        list.AddRange(enclosing.TypeArguments);
-                    }
-                }
+                return false;
             }
 
             /// <summary>
